Commit-2

+++ hw06-atm/src/main/java/ru/otus/BankNote.java
@@ -1,19 +1,19 @@
 package ru.otus;
 
 public enum BankNote {
-    R5000(5000L),
-    R2000(2000L),
-    R1000(1000L),
-    R500(500L),
-    R200(200L),
-    R100(100L);
+    R5000(5000),
+    R2000(2000),
+    R1000(1000),
+    R500(500),
+    R200(200),
+    R100(100);
 
-    private final long rating;
-    BankNote(long rating) {
+    private final int rating;
+    BankNote(int rating) {
         this.rating = rating;
     }
 
-    long getRating(){
+    int getRating(){
         return rating;
     }
